+++ next.config.mjs
@@ -2,10 +2,6 @@ const API_URL = process.env.NEXT_PUBLIC_API_URL;
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: false,
-  async rewrites() {
-    return [
-    ];
-  },
   images: {
     remotePatterns: [
       {
