+++ src/libs/http/refreshToken.ts
@@ -1,6 +1,28 @@
 import axios, { AxiosError } from 'axios'
-import { API_SERVER_URL } from '@/constant'
-import { cookies, setToken } from '../utils/cookie'
+import { cookies } from 'next/headers'
+import { API_SERVER_URL, IS_DEV } from '@/constant'
+
+interface SetTokenParams {
+  accessToken: string
+  refreshToken: string
+}
+
+export async function setToken({ accessToken, refreshToken }: SetTokenParams) {
+  const cookieStore = await cookies()
+  cookieStore.set('access_token', accessToken, {
+    httpOnly: true,
+    maxAge: 60 * 15, // 15分鐘
+    path: '/',
+    secure: !IS_DEV,
+  })
+  cookieStore.set('refresh_token', refreshToken, {
+    httpOnly: true,
+    maxAge: 60 * 60 * 24 * 7, // 7天
+    path: '/',
+    secure: !IS_DEV,
+  })
+}
+
 
 interface RefreshEvent {
   error: AxiosError
